refactor: move ReleaseManageDataBlocksPage components into separate directory
