+++ src/explore-education-statistics-admin/src/pages/release/edit-release/manage-datablocks/components/CreateDataBlocks.tsx
@@ -1,6 +1,6 @@
 import DataBlockDetailsForm, {
   DataBlockDetailsFormValues,
-} from '@admin/pages/release/edit-release/manage-datablocks/DataBlockDetailsForm';
+} from '@admin/pages/release/edit-release/manage-datablocks/components/DataBlockDetailsForm';
 import { mapDataBlockResponseToFullTable } from '@common/modules/find-statistics/components/util/tableUtil';
 import { TableDataQuery } from '@common/modules/table-tool/services/tableBuilderService';
 import { FullTable } from '@common/modules/table-tool/types/fullTable';
